[BUGFIX] Assets had frontend backPath and were 404'd
Some of the assets (JS,CSS) were prefixed with 'typo3/'
URLs which are not suitable when called from Backend pages.
This happened, because RobotsTxtDefault instantiates
TSFE and TSFE instantiates PageRenderer and sets
PageRenderer's backPath to a Frontend backPath.
This patch sets the backPath back to null so that
asset URLs will be rendered in the Backend context
again.

Fixes #150