[TASK] Remove obsolete code for URL generation for AJAX call
The ExtensionManager concept of #147  was already introduced
with #129 via 49958b0.

We now use index.php for AJAX calls via #129 which makes us
not use the deprecation concept (for ajax.php) of #147 anyways.

Plus, do we never determine URLs of AJAX  IDs in the code.
Should we ever need to do that, this commit introduces a
function BackendUtility::getAjaxUrl() which is mostly copied from
the core version 7.5.0-dev (but not used yet).

Issue #129
Fixes #147